infra: fix type conversion in trace
ip6->proto was converted from uint8_t to int to be displayed.
error: format specifies type 'unsigned char' but the argument has type 'int'

+++ modules/infra/datapath/trace.c
@@ -365,7 +365,7 @@ ipv4:
 	case RTE_BE16(RTE_ETHER_TYPE_IPV6): {
 		const struct rte_ipv6_hdr *ip6;
 		uint16_t payload_len;
-		int proto;
+		uint8_t proto;
 
 		ip6 = rte_pktmbuf_mtod_offset(m, const struct rte_ipv6_hdr *, offset);
 		offset += sizeof(*ip6);
